[CDSR-1332][AO] cleanups before adding multiple journey
