+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/views/rejectedgoods/check_claim_details.scala.html
