+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/views/rejectedgoods/check_declaration_details.scala.html
