+++ build.sbt
@@ -34,6 +34,7 @@ lazy val wartremoverSettings =
       Wart.Equals,
       Wart.GlobalExecutionContext,
       Wart.OptionPartial,
+      Wart.EitherProjectionPartial,
       Wart.Throw
     )
   )
