+++ project/plugins.sbt
@@ -8,7 +8,7 @@ addSbtPlugin("uk.gov.hmrc"               % "sbt-auto-build"     % "3.5.0")
 addSbtPlugin("uk.gov.hmrc"               % "sbt-git-versioning" % "2.2.0")
 addSbtPlugin("uk.gov.hmrc"               % "sbt-distributables" % "2.1.0")
 addSbtPlugin("com.typesafe.play"         % "sbt-plugin"         % "2.8.8")
-addSbtPlugin("org.irundaia.sbt"          % "sbt-sassify"        % "1.5.1")
+//addSbtPlugin("org.irundaia.sbt"          % "sbt-sassify"        % "1.5.1")
 addSbtPlugin("org.wartremover"           % "sbt-wartremover"    % "2.4.15")
 addSbtPlugin("org.scalameta"             % "sbt-scalafmt"       % "2.4.0")
 addSbtPlugin("ch.epfl.scala"             % "sbt-scalafix"       % "0.9.29")
