+++ test/uk/gov/hmrc/cdsreimbursementclaimfrontend/journeys/RejectedGoodsSingleJourneyGenerators.scala
@@ -29,7 +29,7 @@ import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.generators._
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.upscan.UploadDocumentType
 
 /** A collection of generators supporting the tests of RejectedGoodsSingleJourney. */
-object RejectedGoodsSingleJourneyGenerators extends RejectedGoodsSingleJourneyTestData {
+object RejectedGoodsSingleJourneyGenerators extends JourneyGenerators with RejectedGoodsSingleJourneyTestData {
 
   val completeJourneyWithMatchingUserEoriAndCMAEligibleGen: Gen[RejectedGoodsSingleJourney] =
     Gen.oneOf(
@@ -113,11 +113,6 @@ object RejectedGoodsSingleJourneyGenerators extends RejectedGoodsSingleJourneyTe
     .submitBasisOfClaim(BasisOfRejectedGoodsClaim.SpecialCircumstances)
     .submitBasisOfClaimSpecialCircumstancesDetails(basisOfClaimSpecialCircumstances)
     .fold(error => throw new Exception(error), identity)
-
-  implicit val bigDecimalChoose = new Gen.Choose[BigDecimal] {
-    override def choose(min: BigDecimal, max: BigDecimal): Gen[BigDecimal] =
-      Gen.choose(1, 10000).map(i => (min + (i * ((max - min) / 10000))).round(min.mc))
-  }
 
   @SuppressWarnings(Array("org.wartremover.warts.Throw"))
   def buildCompleteJourneyGen(
@@ -242,35 +237,6 @@ object RejectedGoodsSingleJourneyGenerators extends RejectedGoodsSingleJourneyTe
             Some(bankAccountType)
           else None
       )
-    }
-
-  val displayDeclarationCMAEligibleGen: Gen[DisplayDeclaration] =
-    buildDisplayDeclarationGen(cmaEligible = true)
-
-  val displayDeclarationNotCMAEligibleGen: Gen[DisplayDeclaration] =
-    buildDisplayDeclarationGen(cmaEligible = false)
-
-  val displayDeclarationGen: Gen[DisplayDeclaration] =
-    Gen.oneOf(
-      displayDeclarationCMAEligibleGen,
-      displayDeclarationNotCMAEligibleGen
-    )
-
-  val exampleDisplayDeclaration: DisplayDeclaration =
-    displayDeclarationGen.sample.get
-
-  def buildDisplayDeclarationGen(cmaEligible: Boolean): Gen[DisplayDeclaration] =
-    for {
-      id               <- Gen.uuid
-      declarantEORI    <- IdGen.genEori
-      consigneeEORI    <- IdGen.genEori
-      numberOfTaxCodes <- Gen.choose(1, 5)
-      taxCodes         <- Gen.const(TaxCodes.all.take(numberOfTaxCodes))
-      paidAmounts      <- Gen.listOfN(numberOfTaxCodes, Gen.choose[BigDecimal](BigDecimal("1.00"), BigDecimal("1000.00")))
-    } yield {
-      val paidDuties: Seq[(TaxCode, BigDecimal, Boolean)] =
-        taxCodes.zip(paidAmounts).map { case (t, a) => (t, a, cmaEligible) }
-      buildDisplayDeclaration(id.toString, declarantEORI, Some(consigneeEORI), paidDuties)
     }
 
 }
